+++ authentication_lib/src/main/java/org/octri/authentication/server/security/service/EmailNotificationService.java
@@ -3,15 +3,12 @@ package org.octri.authentication.server.security.service;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.octri.authentication.config.EmailConfiguration;
 import org.octri.authentication.config.OctriAuthenticationProperties;
 import org.octri.authentication.server.security.AuthenticationUrlHelper;
 import org.octri.authentication.server.security.entity.PasswordResetToken;
 import org.octri.authentication.server.security.entity.User;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.octri.messaging.service.MessageDeliveryService;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
@@ -25,23 +22,38 @@ public class EmailNotificationService {
 
     private static final Log log = LogFactory.getLog(EmailNotificationService.class);
 
-    @Value("${app.displayName}")
-    private String displayName;
+    private final String displayName;
+    private final String fromAddress;
+    private final OctriAuthenticationProperties authenticationProperties;
+    private final AuthenticationUrlHelper urlHelper;
+    private final MessageDeliveryService messageDeliveryService;
+    private final PasswordResetTokenService passwordResetTokenService;
 
-    @Autowired
-    OctriAuthenticationProperties authenticationProperties;
-
-    @Autowired
-    private AuthenticationUrlHelper urlHelper;
-
-    @Autowired
-    private JavaMailSender mailSender;
-
-    @Autowired
-    private EmailConfiguration emailConfig;
-
-    @Autowired
-    private PasswordResetTokenService passwordResetTokenService;
+    /**
+     * Constructor.
+     *
+     * @param displayName
+     *            application display name (injected into messages)
+     * @param authenticationProperties
+     *            configuration properties
+     * @param urlHelper
+     *            authentication URL helper
+     * @param messageDeliveryService
+     *            service used to deliver email messages
+     * @param passwordResetTokenService
+     *            password reset token service
+     */
+    public EmailNotificationService(@Value("${app.displayName}") String displayName,
+            @Value("${spring.mail.from}") String fromAddress, OctriAuthenticationProperties authenticationProperties,
+            AuthenticationUrlHelper urlHelper, MessageDeliveryService messageDeliveryService,
+            PasswordResetTokenService passwordResetTokenService) {
+        this.displayName = displayName;
+        this.fromAddress = fromAddress;
+        this.authenticationProperties = authenticationProperties;
+        this.urlHelper = urlHelper;
+        this.messageDeliveryService = messageDeliveryService;
+        this.passwordResetTokenService = passwordResetTokenService;
+    }
 
     /**
      * Send email confirmation to user. If the user is new, a welcome email is sent. Otherwise a password
@@ -57,28 +69,26 @@ public class EmailNotificationService {
     public void sendPasswordResetTokenEmail(final PasswordResetToken token, final HttpServletRequest request,
             final boolean isNewUser) {
 
-        User user = token.getUser();
+        final User user = token.getUser();
         final String resetPath = urlHelper.getPasswordResetUrl(token.getToken());
-
-        SimpleMailMessage email = new SimpleMailMessage();
+        String subject;
         String body;
+
         if (isNewUser) {
-            email.setSubject("Welcome to " + displayName);
+            subject = "Welcome to " + displayName;
             body = "Hello " + user.getFirstName() + ",\n\nAn account has been created for you with username "
                     + user.getUsername() + ". To set your password, please follow this link: " + resetPath;
 
         } else {
-            email.setSubject("Password reset request for " + displayName);
+            subject = "Password reset request for " + displayName;
             body = "Hello " + user.getFirstName() + ",\n\nTo reset your password please follow this link: "
                     + resetPath + "\n\nIf you did not initiate this request, please contact your system administrator.";
         }
-        email.setText(body);
-        email.setTo(user.getEmail());
-        email.setFrom(emailConfig.getFrom());
+
         if (isEmailDryRunOrBlank(authenticationProperties.getEmailDryRun(), user.getEmail())) {
-            logDryRunEmail(email);
+            logDryRunEmail(fromAddress, user.getEmail(), subject, body);
         } else {
-            mailSender.send(email);
+            messageDeliveryService.sendEmail(fromAddress, user.getEmail(), subject, body);
             log.info("Password reset confirmation email sent to " + user.getEmail());
         }
     }
@@ -100,31 +110,36 @@ public class EmailNotificationService {
 
         final String userEmail = passwordResetToken.getUser().getEmail();
 
-        SimpleMailMessage email = new SimpleMailMessage();
-        email.setSubject("Your " + displayName + " password was reset");
+        final String subject = "Your " + displayName + " password was reset";
         final String body = "Your password has been reset. You may now log into the application.\n\nUsername: "
                 + passwordResetToken.getUser().getUsername()
                 + "\nLink: " + urlHelper.getLoginUrl();
-        email.setText(body);
-        email.setTo(userEmail);
-        email.setFrom(emailConfig.getFrom());
 
         if (isEmailDryRunOrBlank(authenticationProperties.getEmailDryRun(), userEmail)) {
-            logDryRunEmail(email);
+            logDryRunEmail(fromAddress, userEmail, subject, body);
         } else {
-            mailSender.send(email);
+            messageDeliveryService.sendEmail(fromAddress, userEmail, subject, body);
             log.info("Password reset confirmation email sent to " + userEmail);
         }
     }
 
     private static boolean isEmailDryRunOrBlank(final boolean dryRun, final String toEmail) {
+        if (dryRun) {
+            logDeprecationWarning();
+        }
         return dryRun || StringUtils.isBlank(toEmail);
     }
 
-    private static void logDryRunEmail(SimpleMailMessage message) {
+    private static void logDryRunEmail(String fromAddress, String toAddress, String subject, String body) {
         final String format = "DRY RUN, would have sent email to %s from %s with subject \"%s\" and contents \"%s\"";
-        log.info(String.format(format, String.join(", ", message.getTo()), message.getFrom(), message.getSubject(),
-                message.getText()));
+        logDeprecationWarning();
+        log.info(String.format(format, String.join(", ", toAddress), fromAddress, subject,
+                body));
+    }
+
+    private static void logDeprecationWarning() {
+        var message = "Setting octri.authentication.email-dry-run=true is deprecated. Use octri.messaging.email-delivery-strategy=LOG instead.";
+        log.warn(message);
     }
 
 }